+++ src/computelimit/HISTORY.rst
@@ -6,3 +6,5 @@ Release History
 1.0.0b1
 ++++++
 * Initial release.
+* Add shared-limit commands (add, remove, show, list).
+* Add guest-subscription commands (add, remove, show, list).